+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+Remove stripTags filter. Does not work in Magento Community
+Make Matrix go from A to ZZ
 
 ## [1.3.0] - 2017-07-12
 ### Changed