+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 Remove stripTags filter. Does not work in Magento Community
 Make Matrix go from A to ZZ
+Add logic for custom customer address attributes
+Add plugin and observers to save the custom customer address fields
 
 ## [1.3.0] - 2017-07-12
 ### Changed