+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Add logic for custom customer address attributes
 - Add plugin and observers to save the custom customer address fields
 - Remove old street field in forms and use new street fields for saving
+- Prevent error during reorder
 
 ## [1.3.0] - 2017-07-12
 ### Changed