+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Add plugin and observers to save the custom customer address fields
 - Remove old street field in forms and use new street fields for saving
 - Prevent error during reorder
+- Add zipcode validation switch
 
 ## [1.3.0] - 2017-07-12
 ### Changed